Move Project handling logic to a separate file
